+++ simplescn/_client_safe.py
@@ -18,8 +18,8 @@ from simplescn.tools.checks import check_updated_certs, check_local, check_args,
 namestr, hashstr, securitystr, destportint, referencestr, addressstr
 from simplescn._decos import check_args_deco, classify_local, classify_accessable
 
-_checkgetresp =  {"address": str, "port": destportint, "security": securitystr, "traverse_needed": bool}
-_checkgetrespupdate =  {"address": str, "port": destportint, "security": securitystr, "name": namestr, "hash": hashstr, "traverse_needed": bool}
+_checkgetresp =  {"address": addressstr, "security": securitystr, "traverse_needed": bool}
+_checkgetrespupdate =  {"address": addressstr, "security": securitystr, "name": namestr, "hash": hashstr, "traverse_needed": bool}
 
 #@generate_validactions_deco
 class ClientClientSafe(object, metaclass=abc.ABCMeta):